zoho_etl/requirements.txt
requests==2.31.0
google-cloud-bigquery==3.15.0
pandas==2.1.4
python-dotenv==1.0.0
db-dtypes==1.2.0
